backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
clerk-sdk==0.4.2
click==8.3.0
colorama==0.4.6
cryptography==45.0.7
Deprecated==1.3.1
deprecation==2.1.0
fastapi==0.121.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
packaging==25.0
postgrest==2.23.2
propcache==0.4.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
realtime==2.23.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
storage3==2.23.2
StrEnum==0.4.15
supabase==2.23.2
supabase-auth==2.23.2
supabase-functions==2.23.2
svix==1.81.0
types-Deprecated==1.3.1.20251101
types-python-dateutil==2.9.0.20251008
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.0
yarl==1.22.0
